Move OSM-specific functions to `vectorpipe.osm`
